pyproject.toml
[tool.poetry]
name = "teleAgent"
version = "0.1.0"
description = "AI Agent Social Platform"
authors = ["teleAgent"]
readme = "README.md"
packages = [{include = "teleAgent"}]
license = "MIT"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
fastapi = "^0.103.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
redis = "^5.0.0"
injector = "^0.21.0"
python-telegram-bot = "^21.9"
tweepy = "^4.14.0"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
protobuf = "3.20.3"
jinja2 = "^3.1.4"
aiohttp = "^3.11.10"
openai = "^1.3.0" 
requests = "^2.31.0"
arweave-client-python = "^1.0.15.dev0"
anchorpy = "^0.20.1"
solana = ">=0.33.0,<1.0"
solders = "^0.21.0"
typer = "^0.15.1"
autoflake = "^2.3.1"
genpy = "^2022.1"
pytest-xprocess = "^0.23.0"
base58 = "^2.1.1"
pyautogen = { git = "https://github.com/ag2ai/ag2.git", rev = "745a3a5d3cdd2723489e31c9229abe1e089816c2" }
pillow = "^11.0.0"
setuptools = "^75.6.0"
flake8 = "^7.1.1"
replicate = "^1.0.4"
pandas = "^2.2.3"
colorama = "^0.4.6"
tenacity = "^9.0.0"
pyyaml = "^6.0.2"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["teleAgent", "tests"] 
python_files = "*_test.py"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]